Refactor RDF modelling
