+++ spec/models/ore/aggregation_spec.rb
@@ -10,9 +10,9 @@ RSpec.describe ORE::Aggregation do
 
     it { is_expected.to include(Mongoid::Document) }
     it { is_expected.to include(Mongoid::Timestamps) }
-    it { is_expected.to include(Mongoid::Uuid) }
+    it { is_expected.not_to include(Mongoid::Uuid) }
     it { is_expected.to include(Blankness::Mongoid) }
-    it { is_expected.to include(RDFModel) }
+    it { is_expected.to include(RDF::Model) }
 
     it { is_expected.to reject_if_blank(:edm_isShownBy) }
     it { is_expected.to reject_if_blank(:edm_hasViews) }
@@ -20,5 +20,14 @@ RSpec.describe ORE::Aggregation do
 
   it 'should autobuild edm_aggregatedCHO' do
     expect(subject.edm_aggregatedCHO).to be_a(EDM::ProvidedCHO)
+  end
+
+  describe '#rdf_uri' do
+    let(:uuid) { SecureRandom.uuid }
+    subject { build(:ore_aggregation, edm_aggregatedCHO: build(:edm_provided_cho, uuid: uuid)).rdf_uri }
+
+    it 'uses FQDN, /stories and UUID of aggregations' do
+      expect(subject).to eq(RDF::URI.new("https://stories.europeana.eu/stories/#{uuid}"))
+    end
   end
 end
