+++ spec/models/story_spec.rb
@@ -7,6 +7,7 @@ RSpec.describe Story do
     subject { described_class }
     it { is_expected.to include(Mongoid::Document) }
     it { is_expected.to include(Mongoid::Timestamps) }
+    it { is_expected.to include(RDF::Model::Dumping) }
   end
 
   it 'should autobuild ore_aggregation' do
